+++ yast/response.py
@@ -1,7 +1,12 @@
-from .datastructures import MutableHeaders
-from .types import Recevie, Send
+import aiofiles
 import json
 import typing
+
+from mimetypes import guess_type
+from yast.types import Recevie, Send
+from yast.datastructures import MutableHeaders
+from yast.types import Recevie, Send
+
 
 
 class Response:
@@ -19,7 +24,7 @@ class Response:
         self.status_code = status_code
         if media_type is not None:
             self.media_type = media_type
-        self.set_default_headers(headers)
+        self.init_headers(headers)
 
     async def __call__(self, receive: Recevie, send: Send) -> None:
         await send(
@@ -38,7 +43,7 @@ class Response:
             return content
         return content.encode(self.charset)
 
-    def set_default_headers(self, headers: dict = None):
+    def init_headers(self, headers: dict = None):
         if headers is None:
             raw_headers = []
             missing_content_length = True
@@ -48,17 +53,18 @@ class Response:
                 (h_k.lower().encode('latin-1'), h_v.encode('latin-1'))
                 for h_k, h_v in headers.items()
             ]
-            missing_content_length = 'content-length' not in headers
-            missing_content_type = 'content-type' not in headers
+            keys = [h[0] for h in raw_headers]
+            missing_content_length = b'content-length' not in keys
+            missing_content_type = b'content-type' not in keys
+
+        body = getattr(self, 'body', None)
+        if body is not None and missing_content_length:
+            raw_headers.append((b'content-length', str(len(body)).encode('latin-1')))
         
-        if missing_content_length:
-            raw_headers.append((b'content-length', str(len(self.body)).encode()))
-        
-        if self.media_type is not None and missing_content_type:
-            content_type = self.media_type
-            if content_type.startswith('text/') and self.charset is not None:
+        content_type = self.media_type
+        if content_type is not None and missing_content_type:
+            if content_type.startswith('text/'):
                 content_type += '; charset=%s' % self.charset
-            
             raw_headers.append((b'content-type', content_type.encode('latin-1')))
 
         self.raw_headers = raw_headers
@@ -99,7 +105,7 @@ class StreamingResponse(Response):
         if media_type:
             self.media_type = media_type
 
-        self.set_default_headers(headers)
+        self.init_headers(headers)
 
     async def __call__(self, receive: Recevie, send: Send):
         await send(
@@ -118,22 +124,46 @@ class StreamingResponse(Response):
             await send({ "type": "http.response.body", "body": chunk, "more_body": True })
         await send({ "type": "http.response.body", "body": b"", "more_body": False })
 
-    def set_default_headers(self, headers: dict = None):
-        if headers is None:
-            raw_headers = []
-            missing_content_type = True
-        else:
-            raw_headers = [
-                (h_k.lower().encode('latin-1'), h_v.encode('latin-1'))
-                for h_k, h_v in headers.items()
-            ]
-            missing_content_type = 'content-type' not in headers
-        
-        if self.media_type is not None and missing_content_type:
-            content_type = self.media_type
-            if content_type.startswith('text/') and self.charset is not None:
-                content_type += '; charset=%s' % self.charset
-            
-            raw_headers.append((b'content-type', content_type.encode('latin-1')))
 
-        self.raw_headers = raw_headers
+class FileResponse(Response):
+    """ File Response """
+    chunk_size = 4096
+
+    def __init__(
+            self,
+            path: str,
+            headers: dict = None,
+            media_type: str = None,
+            filename: str = None
+        ) -> None:
+        """ init """
+        self.path = path
+        self.status_code = 200
+        self.filename = filename
+        if media_type is None:
+            media_type = guess_type(filename or path)[0] or 'text/plain'
+        
+        self.media_type = media_type
+        self.init_headers(headers)
+        if self.filename is not None:
+            content_disposition = f'attachment; filename="{self.filename}"'
+            self.headers.setdefault('content-disposition', content_disposition)
+    
+
+    async def __call__(self, receive: Recevie, send: Send) -> None:
+        await send({
+            'type': 'http.response.start',
+            'status': self.status_code,
+            'headers': self.raw_headers,
+        })
+
+        async with aiofiles.open(self.path, mode='rb') as file:
+            more_body = True
+            while more_body:
+                chunk = await file.read(self.chunk_size)
+                more_body = len(chunk) == self.chunk_size
+                await send({
+                    'type': 'http.response.body',
+                    'body': chunk,
+                    'more_body': False, # todo: why
+                })